llms/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
asttokens==3.0.0
attrs==24.3.0
boto3==1.36.1
botocore==1.36.1
certifi==2024.12.14
charset-normalizer==3.4.1
comm==0.2.2
datasets==3.2.0
debugpy==1.8.12
decorator==5.1.1
dill==0.3.8
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kaleido==0.2.1
matplotlib-inline==0.1.7
multidict==6.1.0
multiprocess==0.70.16
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.2.1
ollama==0.4.7
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.48
propcache==0.2.1
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.1
requests==2.32.3
rpds-py==0.22.3
s3transfer==0.11.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tenacity==9.0.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.18.3
